celebrium-docker-prod/requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy
torch
torchvision
python-multipart